LWeapon Scripts: Some Experiments : WTF?
I'm doing some experiments with placement.

For some reason, the weapon script ID (value assigned to w->weaponscript)
is not always correct, or the function in FFCore that is reading it can't
figure it out.

If I launch a swordbeam lweapon, some times it does, and some times it does not
properly read its script; but a boomerang lweapon always does.

Something's screwy.

Perhaps there is a var that's mis-initialised. I'm usure.

I'd love to run this on valgrind, but I need to do that on Linux, and I don't
have the spare system capacity to leave this running forever on LX just for that.